+++ packages/terra-search-field/src/terra-dev-site/doc/example/SearchFieldBasic.jsx
@@ -2,15 +2,10 @@ import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import SearchField from 'terra-search-field/lib/SearchField';
 
-const defaultProps = {
-  minimumSearchTextLength: 2,
-};
-
 class SearchFieldBasic extends React.Component {
   constructor(props) {
     super(props);
-    this.defaultLength = this.props.minimumSearchTextLength === 2 ? 'default' : '';
-    this.invalidSearchMessage = `The ${this.defaultLength} minimum search length is ${this.props.minimumSearchTextLength}.`;
+    this.invalidSearchMessage = 'The defualt minimum search length is 2.';
     this.state = { searchText: '', message: this.invalidSearchMessage };
     this.handleSearch = this.handleSearch.bind(this);
     this.handleInvalidSearch = this.handleInvalidSearch.bind(this);
@@ -25,7 +20,6 @@ class SearchFieldBasic extends React.Component {
   }
 
   render() {
-
     return (
       <div>
         <p>
@@ -40,7 +34,5 @@ class SearchFieldBasic extends React.Component {
     );
   }
 }
-
-SearchFieldBasic.defaultProps = defaultProps;
 
 export default SearchFieldBasic;