+++ packages/terra-search-field/src/terra-dev-site/doc/example/SearchFieldDisableAutoFormatAttrs.jsx
@@ -2,15 +2,10 @@ import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import SearchField from 'terra-search-field/lib/SearchField';
 
-const defaultProps = {
-  minimumSearchTextLength: 2,
-};
-
 class SearchFieldDisableAutoFormatAttrs extends React.Component {
   constructor(props) {
     super(props);
-    this.defaultLength = this.props.minimumSearchTextLength === 2 ? 'default' : '';
-    this.invalidSearchMessage = `The ${this.defaultLength} minimum search length is ${this.props.minimumSearchTextLength}.`;
+    this.invalidSearchMessage = 'The default minimum search length is 2.';
     this.state = { searchText: '', message: this.invalidSearchMessage };
     this.handleSearch = this.handleSearch.bind(this);
     this.handleInvalidSearch = this.handleInvalidSearch.bind(this);
@@ -25,7 +20,6 @@ class SearchFieldDisableAutoFormatAttrs extends React.Component {
   }
 
   render() {
-
     return (
       <div>
         <p>
@@ -33,12 +27,12 @@ class SearchFieldDisableAutoFormatAttrs extends React.Component {
           {this.state.searchText}
         </p>
         <SearchField
-         inputAttributes={{
-          autoCorrect: 'off',
-          autoComplete: 'off',
-          autoCapitalize: 'off',
-          spellCheck: 'false',
-        }}
+          inputAttributes={{
+            autoCorrect: 'off',
+            autoComplete: 'off',
+            autoCapitalize: 'off',
+            spellCheck: 'false',
+          }}
           onSearch={this.handleSearch}
           onInvalidSearch={this.handleInvalidSearch}
         />
@@ -47,7 +41,4 @@ class SearchFieldDisableAutoFormatAttrs extends React.Component {
   }
 }
 
-SearchFieldDisableAutoFormatAttrs.defaultProps = defaultProps;
-
 export default SearchFieldDisableAutoFormatAttrs;
-